Fix save_fs flag argument confusion
The names of the two flags in save_fs were the exact opposite
of their names in BundleFile fields and BundleFile.save.
In particular, bundle_file.save("original") would end up swapping
the flags, which not only was counterintuitive but could even make
the output files unreadable in Unity proper (because most
of the flags aren't validated or enforced by UnityPy).

Rename the save_fs arguments to match BundleFile fields.
Keep them in the same order so that any user-written calls
without named arguments continue to work.