Breaking Change: Fix objects with the same container path
Reimplement the array of KeyValuePair (multi-dict) structure in AssetStudio as a list-of-tuple structure to address the issue of multiple objects with the same container path when implemented as dict.

Breaking Part: All .items() call to .container or .m_Container is no longer needed as it is an array instead of a dict.